Move merch types to data/types.ts
